:bug: fix: in 'CheckPawnsAhead()' there was a missing conditional, to check if 'tAhead.Value.Count' was greater than 'pos' before trying to use the method 'RemoveRange()'

+++ Classes/Automation/Strategies/LastMan.cs
@@ -117,8 +117,11 @@ namespace CartagenaBuenaventura.Classes.Automation.Strategies
                     {
                         if (ahead.Find(x => x.position == pos) == null)
                         {
-                            tAhead.Value.RemoveRange(pos, tAhead.Value.Count);
-                            break;
+                            if (tAhead.Value.Count > pos) 
+                            {
+                                tAhead.Value.RemoveRange(pos, tAhead.Value.Count);
+                                break;
+                            }
                         }
                     }
                 }
